Split org components into multiple files
