+++ src/components/Organizations/Details/Elections.tsx
@@ -1,0 +1,65 @@
+import { Flex, Text } from '@chakra-ui/react'
+import { AccountData, PublishedElection } from '@vocdoni/sdk'
+import { Trans } from 'react-i18next'
+import { useParams } from 'react-router-dom'
+import { LoadingCards } from '~components/Layout/Loading'
+import { RoutedPagination } from '~components/Pagination/Pagination'
+import { RoutedPaginationProvider } from '~components/Pagination/PaginationProvider'
+import { ElectionCard } from '~components/Process/Card'
+import { PaginationItemsPerPage, RoutePath } from '~constants'
+import { useOrganizationElections } from '~queries/organizations'
+import { retryUnlessNotFound } from '~utils/queries'
+
+interface OrgComponentProps {
+  org: AccountData
+}
+
+const OrganizationElections = ({ org }: OrgComponentProps) => {
+  if (org.electionIndex === 0) {
+    return (
+      <Text>
+        <Trans i18nKey={'organization.no_elections'}>No elections yet!</Trans>
+      </Text>
+    )
+  }
+
+  return (
+    <RoutedPaginationProvider
+      totalPages={Math.ceil(org.electionIndex / PaginationItemsPerPage)}
+      path={RoutePath.Organization}
+    >
+      <Flex direction={'column'} gap={4}>
+        <OrganizationElectionsList org={org} />
+        <RoutedPagination />
+      </Flex>
+    </RoutedPaginationProvider>
+  )
+}
+
+const OrganizationElectionsList = ({ org }: OrgComponentProps) => {
+  const { page } = useParams()
+
+  const { data: elections, isLoading } = useOrganizationElections({
+    address: org.address,
+    page: Number(page) - 1 || 0,
+    options: {
+      enabled: !!org.address,
+      retry: retryUnlessNotFound,
+    },
+  })
+
+  if (isLoading) {
+    return <LoadingCards />
+  }
+
+  return (
+    <Flex direction={'column'} gap={4}>
+      {elections?.map((election) => {
+        if (election instanceof PublishedElection) return <ElectionCard key={election.id} election={election} />
+        return null
+      })}
+    </Flex>
+  )
+}
+
+export default OrganizationElections
